full implementation for incremental tiger index

+++ tools/server/storage-engine/ColumnarTiger_v1.hpp
@@ -1274,7 +1274,7 @@ public:
          // }
 
          for (auto& p : new_map) {
-             auto existing_rows = idx_engine->LookupIndex("term", p.first, increment);
+             auto existing_rows = idx_engine->LookupIndex("term", p.first, (int)increment);
              if (existing_rows.empty()) {
                  // New term
                  RowID r = idx_engine->AddRow();
